backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
supabase
chardet
rfc3986
httpx
fastapi
uvicorn[standard]
pydantic
pydantic-settings
supabase
python-dotenv

passlib[bcrypt]
python-jose[cryptography]
bcrypt===4.0.1
